Updated SteamConsole Update Script
Changed RetroArch nightly download to detect yesterdays date and try
downloading with that instead of today's date. There are some days
RetroArch doesn't have a build for the current day so I felt it was
better to try for yesterday since it will typically be there in more
situations than today.